Add overlay, close on clicking outside, design improvements

+++ src/App.jsx
@@ -1,16 +1,20 @@
 import React, { useState, useEffect } from "react";
 import "./App.css";
-import CardsList from "./components/CardsList";
 import AddNotePopup from "./components/AddNotePopup";
+import CardsList from "./components/CardsList";
 import { nanoid } from "nanoid";
 import axios from "axios";
 import { ToastContainer } from "react-toastify";
 import Search from "./components/Search";
 import Header from "./components/Header";
+import { toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.css";
 
 //API Details
 // 20220912223415
 // https://6315b6ef33e540a6d38296a9.mockapi.io/notepad-app
+
+const ID = nanoid();
 
 function App() {
   const randomColor1 = Math.floor(Math.random() * 16777215).toString(16);
@@ -18,7 +22,7 @@ function App() {
 
   const [notes, setNotes] = useState([
     {
-      id: nanoid(),
+      id: ID,
       color1: randomColor1,
       color2: randomColor2,
       text: "this is the note pad app text",
@@ -49,7 +53,7 @@ function App() {
   const addNote = (text) => {
     const date = new Date();
     const newNote = {
-      id: nanoid(),
+      id: ID,
       color1: randomColor1,
       color2: randomColor2,
       text: text,
@@ -69,24 +73,49 @@ function App() {
   const deleteNote = (id) => {
     const newNotes = notes.filter((note) => note.id !== id);
     //delete operation
-    axios.delete(
-      `https://6315b6ef33e540a6d38296a9.mockapi.io/notepad-app/${id}`,
-      setNotes(newNotes)
-    );
+    if (id == ID)
+      axios.delete(
+        `https://6315b6ef33e540a6d38296a9.mockapi.io/notepad-app/${id}`,
+        setNotes(newNotes)
+      );
+    else
+      toast("📋 This is not YourQutoes", {
+        position: "top-center",
+        autoClose: 2000,
+        hideProgressBar: false,
+        closeOnClick: true,
+        pauseOnHover: true,
+        draggable: true,
+        progress: undefined,
+      });
+  };
+
+  const checkIfClickedInside = (e) => {
+    if (
+      e.target.dataset.target != "popup" &&
+      e.target.dataset.target != "add-quote"
+    ) {
+      setAddNotePopupIsOpen(false);
+    }
   };
 
   return (
     <>
-      <div className={`${darkMode && "dark-mode"}`}>
+      <div
+        className={`${darkMode && "dark-mode"}`}
+        onClick={checkIfClickedInside}
+      >
         <Header handleToggleDarkMode={setDarkMode} />
-        <div className="container">
+        <div className={`container ${addNotePopupIsOpen && "add-overlay"}`}>
+          <div className="wrapper"></div>
           <Search handleSearchNote={setSearchText} />
           <CardsList
-            notes={notes.filter((note) => note.text.includes(searchText))}
+            notes={notes.filter((note) =>
+              note.text.toUpperCase().includes(searchText.toLocaleUpperCase())
+            )}
             handleAddNote={addNote}
             handleDeleteNote={deleteNote}
           />
-          {/* <AddNotePopup /> */}
           <ToastContainer
             position="top-center"
             autoClose={2000}
@@ -104,16 +133,17 @@ function App() {
             onClick={() => {
               setAddNotePopupIsOpen(true);
             }}
+            data-target="add-quote"
           >
-            <i class="fa-solid fa-plus"></i>
+            <i class="fa-solid fa-plus" data-target="add-quote"></i>
           </button>
-          {addNotePopupIsOpen && (
-            <AddNotePopup
-              handleAddNote={addNote}
-              setAddNotePopupIsOpen={setAddNotePopupIsOpen}
-            />
-          )}
         </div>
+        {addNotePopupIsOpen && (
+          <AddNotePopup
+            handleAddNote={addNote}
+            setAddNotePopupIsOpen={setAddNotePopupIsOpen}
+          />
+        )}
       </div>
     </>
   );
